Fix `AttachConsole` when STDOUT is already attached (#38)
This fixes a bug with `AttachConsole` when STDOUT is already attached and available. In that case, we'd attach a console, and re-allocate stdout, which would break existing pipes and cause a crash.

In the case where `stdout` is already available, don't bother calling the `AttachConsole` API - we already have a `stdout` to use.